Enable docker-init for the testsuite container
This goes with a change in the corresponding image, to not install tini,
since it now comes by default with docker and is available from
docker-compose with a recent syntax.
Moreover add a command to sleep infinitely, waiting for an external
command.

See: issue https://issues.infn.it/jira/browse/VOMS-882